link-analyzer/requirements.txt
# FastAPI core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP client
httpx==0.26.0

# Redis for caching
redis==5.0.1
hiredis==2.3.2

# AI/ML - Gemini 2.5 API
google-genai==0.3.0  # Latest Gemini SDK with video support
google-generativeai==0.3.2  # Legacy SDK (fallback)

# Google APIs
google-api-python-client==2.110.0  # YouTube Data API v3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Video/Media Download
yt-dlp==2024.1.1  # Universal video downloader

# Database
sqlalchemy==2.0.25  # ORM for PostgreSQL
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
